requirements_test.txt
# Strictly for tests
pytest==7.4.4
#pytest-cov==2.9.0
#pytest-homeassistant
pytest-homeassistant-custom-component==0.13.103
# From our manifest.json for our custom component
pymelcloud==2.11.0
